+++ protocols/src/main/scala/dx/util/protocols/S3FileAccessProtocol.scala
@@ -18,6 +18,7 @@ import dx.util.{
   AbstractAddressableFileNode,
   AddressableFileSource,
   FileAccessProtocol,
+  FileSource,
   FileUtils,
   Logger
 }
@@ -47,6 +48,11 @@ case class S3FileSource(
     case null => ""
     case path => path.toString
   }
+
+  override def container: String = s"${S3FileAccessProtocol.S3Scheme}:${bucketName}:${folder}"
+
+  // TODO: handle version
+  override def version: Option[String] = None
 
   override def exists: Boolean = {
     try {
@@ -97,6 +103,8 @@ case class S3FolderSource(override val address: String, bucketName: String, pref
     case null => ""
     case path => path.toString
   }
+
+  override def container: String = s"${S3FileAccessProtocol.S3Scheme}:${bucketName}:${folder}"
 
   override val isDirectory: Boolean = true
 
@@ -151,6 +159,25 @@ case class S3FolderSource(override val address: String, bucketName: String, pref
                                    ResponseTransformer.toFile[GetObjectResponse](destPath))
     }
   }
+
+  override def listing: Vector[FileSource] = {
+    val sourcePath = Paths.get(prefix)
+    val relPaths = listPrefix.map(s3obj => sourcePath.relativize(Paths.get(s3obj.key())) -> s3obj)
+    val files = relPaths.collect {
+      case (relPath, s3Obj) if relPath.getNameCount == 1 =>
+        S3FileSource(s"s3://${bucketName}/${s3Obj.key()}", bucketName, s3Obj.key())(protocol)
+    }
+    val folders = relPaths
+      .collect {
+        case (relPath, _) if relPath.getNameCount == 2 =>
+          sourcePath.resolve(relPath.getParent).toString
+      }
+      .toSet
+      .map { folder: String =>
+        S3FolderSource(s"s3://${bucketName}/${folder}", bucketName, folder)(protocol)
+      }
+    files ++ folders.toVector
+  }
 }
 
 case class S3FileAccessProtocol(region: Region,
@@ -166,10 +193,12 @@ case class S3FileAccessProtocol(region: Region,
     client.get
   }
 
-  override val schemes: Vector[String] = Vector("s3")
+  override val schemes: Vector[String] = Vector(S3FileAccessProtocol.S3Scheme)
 
   override val supportsDirectories: Boolean = true
 
+  // TODO: handle version
+  //  how is version specified (I think there is a ?version=<version>)
   private val S3UriRegexp = "s3://(.+?)/(.*)".r
 
   /**
@@ -221,6 +250,7 @@ case class S3FileAccessProtocol(region: Region,
 }
 
 object S3FileAccessProtocol {
+  val S3Scheme = "s3"
 
   /**
     * Creates a S3FileAccessProtocol for the AWS region corresponding to the given