+++ api/RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # dxApi
 
+## 0.3.0 (2021-05-07)
+
+* Adds functions to create `dx://` URIs from components
+* URL-encodes project names, paths, and file names when creating `dx://` URIs, and decodes them when parsing 
+* Adds `state` parameter to `DxFindDataObjects`
+* Adds `DxApi.uploadDirectory` function
+
 ## 0.2.0 (2021-04-23)
 
 * Adds option to wait for upload to `DxApi.uploadFile`